consolidate code; only support go1.10+ (the go mod file requres go1.17)
